# Conflicts:
#	.gitignore
#	analyzer/windows/analyzer.py
#	analyzer/windows/modules/packages/dll.py
#	conf/api.conf
#	conf/auxiliary.conf
#	conf/cuckoomx.conf
#	conf/esx.conf
#	conf/kvm.conf
#	conf/kvmremote.conf
#	conf/malheur.conf
#	conf/physical.conf
#	conf/processing.conf
#	conf/proxmox.conf
#	conf/qemu.conf
#	conf/reporting.conf
#	conf/smtp_sinkhole.conf
#	conf/virtualbox.conf
#	conf/vmware.conf
#	conf/vmwareserver.conf
#	conf/vpn.conf
#	conf/vsphere.conf
#	conf/web.conf
#	conf/xenserver.conf
#	data/safelist/domains.py
#	extra/whitelist_domains.txt
#	lib/cuckoo/common/abstracts.py
#	lib/cuckoo/common/web_utils.py
#	lib/cuckoo/core/database.py
#	lib/cuckoo/core/plugins.py
#	lib/cuckoo/core/resultserver.py
#	lib/cuckoo/core/scheduler.py
#	modules/processing/behavior.py
#	modules/processing/network.py
#	modules/processing/static.py
#	modules/reporting/litereport.py
#	modules/reporting/submitCAPE.py
#	modules/signatures/CAPE.py
#	requirements.txt
#	systemd/cape-web.service
#	utils/cleaners.py
#	utils/community.py
#	utils/process.py
#	utils/route.py
#	web/apiv2/views.py

+++ analyzer/linux/modules/auxiliary/stap.py
@@ -2,15 +2,13 @@
 # This file is part of Cuckoo Sandbox - http://www.cuckoosandbox.org
 # See the file 'docs/LICENSE' for copying permission.
 
-from __future__ import absolute_import
 import logging
 import os
 import subprocess
-import time
+import timeit
 
 from lib.common.abstracts import Auxiliary
 from lib.common.results import upload_to_host
-from lib.core.config import Config
 
 log = logging.getLogger(__name__)
 
@@ -20,48 +18,70 @@ class STAP(Auxiliary):
 
     priority = -10  # low prio to wrap tightly around the analysis
 
-    def __init__(self, options={}, analyzer=None):
-        self.config = Config(cfg="analysis.conf")
+    def __init__(self, options, config):
+        Auxiliary.__init__(self, options, config)
+        self.config = config
+        self.enabled = self.config.stap
         self.proc = None
 
     def start(self):
+        if not self.enabled:
+            return False
+
         # helper function locating the stap module
         def has_stap(p):
-            only_stap = [fn for fn in os.listdir(p) if fn.startswith("stap_") and fn.endswith(".ko")]
-            if only_stap:
-                return os.path.join(p, only_stap[0])
+            for fn in os.listdir(p):
+                if fn.startswith("stap_") and fn.endswith(".ko"):
+                    return os.path.join(p, fn)
             return False
 
-        path_cfg = self.config.get("analyzer_stap_path", None)
+        path_cfg = self.config.get("analyzer_stap_path")
+        cuckoo_path = os.path.join("/root", ".cuckoo")
+        cape_path = os.path.join("/root", ".cape")
         if path_cfg and os.path.exists(path_cfg):
             path = path_cfg
-        elif os.path.exists("/root/.cuckoo") and has_stap("/root/.cuckoo"):
-            path = has_stap("/root/.cuckoo")
-        elif os.path.exists("/root/.cape") and has_stap("/root/.cape"):
-            path = has_stap("root/.cape")
+        elif os.path.exists(cuckoo_path) and has_stap(cuckoo_path):
+            path = has_stap(cuckoo_path)
+        elif os.path.exists(cape_path) and has_stap(cape_path):
+            path = has_stap(cape_path)
         else:
-            log.warning("Could not find STAP LKM, aborting systemtap analysis.")
+            log.warning("Could not find STAP LKM, aborting systemtap analysis")
             return False
 
-        stap_start = time.time()
-        self.proc = subprocess.Popen(["staprun", "-vv", "-x", str(os.getpid()), "-o", "stap.log", path,], stderr=subprocess.PIPE)
+        stap_start = timeit.default_timer()
+        self.proc = subprocess.Popen(
+            [
+                "staprun",
+                "-vv",
+                "-x",
+                str(os.getpid()),
+                "-o",
+                "stap.log",
+                path,
+            ],
+            stderr=subprocess.PIPE,
+        )
 
-        while "systemtap_module_init() returned 0" not in self.proc.stderr.readline().decode("utf8"):
+        while b"systemtap_module_init() returned 0" not in self.proc.stderr.readline():
             pass
 
         self.proc.terminate()
         self.proc.wait()
 
-        stap_stop = time.time()
-        log.info("STAP aux module startup took %.2f seconds" % (stap_stop - stap_start))
+        stap_stop = timeit.default_timer()
+        log.info("STAP aux module startup took %.2f seconds", stap_stop - stap_start)
+
         return True
 
     def stop(self):
+        if not self.enabled:
+            return False
+
         try:
             r = self.proc.poll()
-            log.debug("stap subprocess retval %r", r)
+            log.debug("stap subprocess retval %d", r)
             self.proc.kill()
         except Exception as e:
             log.warning("Exception killing stap: %s", e)
 
-        upload_to_host("stap.log", "stap/stap.log", False)
+        upload_to_host("stap.log", "stap/stap.log", True)
